separate out some utils
