+++ src/artemis/experiment_plans/oav_centring_plan.py
@@ -10,15 +10,14 @@ from dodal.devices.oav.oav_calculations import (
     get_rotation_increment,
     keep_inside_bounds,
 )
-from dodal.devices.oav.oav_detector import OAV, ColorMode, EdgeOutputArrayImageType
-from dodal.devices.oav.oav_errors import (
-    OAVError_WaveformAllZero,
-    OAVError_ZoomLevelNotFound,
-)
+from dodal.devices.oav.oav_detector import OAV
+from dodal.devices.oav.oav_errors import OAVError_WaveformAllZero
 from dodal.devices.oav.oav_parameters import OAVParameters
 from dodal.devices.smargon import Smargon
 
+from artemis.device_setup_plans.setup_oav import pre_centring_setup_oav
 from artemis.log import LOGGER, set_up_logging_handlers
+from artemis.utils.oav_utils import get_waveforms_to_image_scale
 
 # Z and Y bounds are hardcoded into GDA (we don't want to exceed them). We should look
 # at streamlining this
@@ -29,107 +28,6 @@ _Y_UPPER_BOUND = _Z_UPPER_BOUND = 1500
 # reflect this. Despite this, Neil would like to have omega to oscillate so we will
 # hard code limits so gridscans will switch rotation directions and |omega| will stay pretty low.
 _DESIRED_HIGH_LIMIT = 181
-
-
-def start_mxsc(oav: OAV, input_plugin, min_callback_time, filename):
-    """
-    Sets PVs relevant to edge detection plugin.
-
-    Args:
-        input_plugin: link to the camera stream
-        min_callback_time: the value to set the minimum callback time to
-        filename: filename of the python script to detect edge waveforms from camera stream.
-    Returns: None
-    """
-    yield from bps.abs_set(oav.mxsc.input_plugin_pv, input_plugin)
-
-    # Turns the area detector plugin on
-    yield from bps.abs_set(oav.mxsc.enable_callbacks_pv, 1)
-
-    # Set the minimum time between updates of the plugin
-    yield from bps.abs_set(oav.mxsc.min_callback_time_pv, min_callback_time)
-
-    # Stop the plugin from blocking the IOC and hogging all the CPU
-    yield from bps.abs_set(oav.mxsc.blocking_callbacks_pv, 0)
-
-    # Set the python file to use for calculating the edge waveforms
-    current_filename = yield from bps.rd(oav.mxsc.py_filename)
-    if current_filename != filename:
-        LOGGER.info(f"Current python file is {current_filename}, setting to {filename}")
-        yield from bps.abs_set(oav.mxsc.py_filename, filename)
-        yield from bps.abs_set(oav.mxsc.read_file, 1)
-
-    # Image annotations
-    yield from bps.abs_set(oav.mxsc.draw_tip, True)
-    yield from bps.abs_set(oav.mxsc.draw_edges, True)
-
-    # Use the original image type for the edge output array
-    yield from bps.abs_set(oav.mxsc.output_array, EdgeOutputArrayImageType.ORIGINAL)
-
-
-def pre_centring_setup_oav(oav: OAV, parameters: OAVParameters):
-    """Setup OAV PVs with required values."""
-
-    parameters.load_parameters_from_json()
-
-    yield from bps.abs_set(oav.cam.color_mode, ColorMode.RGB1)
-    yield from bps.abs_set(oav.cam.acquire_period, parameters.acquire_period)
-    yield from bps.abs_set(oav.cam.acquire_time, parameters.exposure)
-    yield from bps.abs_set(oav.cam.gain, parameters.gain)
-
-    # select which blur to apply to image
-    yield from bps.abs_set(oav.mxsc.preprocess_operation, parameters.preprocess)
-
-    # sets length scale for blurring
-    yield from bps.abs_set(oav.mxsc.preprocess_ksize, parameters.preprocess_K_size)
-
-    # Canny edge detect
-    yield from bps.abs_set(
-        oav.mxsc.canny_lower_threshold,
-        parameters.canny_edge_lower_threshold,
-    )
-    yield from bps.abs_set(
-        oav.mxsc.canny_upper_threshold,
-        parameters.canny_edge_upper_threshold,
-    )
-    # "Close" morphological operation
-    yield from bps.abs_set(oav.mxsc.close_ksize, parameters.close_ksize)
-
-    # Sample detection
-    yield from bps.abs_set(
-        oav.mxsc.sample_detection_scan_direction, parameters.direction
-    )
-    yield from bps.abs_set(
-        oav.mxsc.sample_detection_min_tip_height,
-        parameters.minimum_height,
-    )
-
-    # Connect MXSC output to MJPG input
-    yield from start_mxsc(
-        oav,
-        parameters.input_plugin + "." + parameters.mxsc_input,
-        parameters.min_callback_time,
-        parameters.filename,
-    )
-
-    yield from bps.abs_set(oav.snapshot.input_pv, parameters.input_plugin + ".CAM")
-
-    zoom_level_str = f"{float(parameters.zoom)}x"
-    if zoom_level_str not in oav.zoom_controller.allowed_zoom_levels:
-        raise OAVError_ZoomLevelNotFound(
-            f"Found {zoom_level_str} as a zoom level but expected one of {oav.zoom_controller.allowed_zoom_levels}"
-        )
-
-    yield from bps.abs_set(
-        oav.zoom_controller.level,
-        zoom_level_str,
-        wait=True,
-    )
-    yield from bps.wait()
-
-    """
-    TODO: We require setting the backlight brightness to that in the json, we can't do this currently without a PV.
-    """
 
 
 def rotate_pin_and_collect_positional_data(
@@ -204,23 +102,6 @@ def rotate_pin_and_collect_positional_data(
         tip_i_positions,
         tip_j_positions,
     )
-
-
-def get_waveforms_to_image_scale(oav: OAV):
-    """
-    Returns the scale of the image. The standard calculation for the image is based
-    on a size of (1024, 768) so we require these scaling factors.
-    Args:
-        oav (OAV): The OAV device in use.
-    Returns:
-        The (i_dimensions,j_dimensions) where n_dimensions is the scale of the camera image to the
-        waveform values on the n axis.
-    """
-    image_size_i = yield from bps.rd(oav.cam.array_size.array_size_x)
-    image_size_j = yield from bps.rd(oav.cam.array_size.array_size_y)
-    waveform_size_i = yield from bps.rd(oav.mxsc.waveform_size_x)
-    waveform_size_j = yield from bps.rd(oav.mxsc.waveform_size_y)
-    return image_size_i / waveform_size_i, image_size_j / waveform_size_j
 
 
 def centring_plan(
