+++ src/artemis/unit_tests/test_fast_grid_scan_plan.py
@@ -40,7 +40,7 @@ from artemis.external_interaction.system_tests.conftest import (
 from artemis.log import set_up_logging_handlers
 from artemis.parameters.external_parameters import RawParameters
 from artemis.parameters.internal_parameters import InternalParameters
-from artemis.utils import Point3D
+from artemis.utils.utils import Point3D
 
 
 @pytest.fixture
